Fix gradle.properties for incubator test Move eqk and genk from ioExtensions to its correspondent util class Fix imports and so after fixing arrow core test and fx test
