+++ arrow-incubator-test/gradle.properties
@@ -1,5 +1,5 @@
 # Maven publishing configuration
-POM_NAME=Arrow-Test
-POM_ARTIFACT_ID=arrow-test
+POM_NAME=Arrow-Incubator-Test
+POM_ARTIFACT_ID=arrow-incubator-test
 POM_PACKAGING=jar
 kapt.incremental.apt=false
