+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/AccumTTest.kt
@@ -19,11 +19,19 @@ import arrow.core.extensions.option.alternative.alternative
 import arrow.core.extensions.option.eqK.eqK
 import arrow.core.extensions.option.monad.monad
 import arrow.core.extensions.tuple2.eq.eq
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.GenK
+import arrow.core.test.generators.genK
+import arrow.core.test.generators.throwable
+import arrow.core.test.generators.tuple2
+import arrow.core.test.laws.AlternativeLaws
+import arrow.core.test.laws.MonadErrorLaws
 import arrow.core.toT
 import arrow.fx.IO
 import arrow.fx.extensions.io.monadIO.monadIO
 import arrow.fx.fix
 import arrow.fx.mtl.accumt.monadIO.monadIO
+import arrow.fx.test.laws.equalUnderTheLaw
 import arrow.mtl.extensions.accumt.alternative.alternative
 import arrow.mtl.extensions.accumt.functor.functor
 import arrow.mtl.extensions.accumt.monad.monad
@@ -36,14 +44,6 @@ import arrow.mtl.extensions.statet.monadState.monadState
 import arrow.mtl.extensions.writert.eqK.eqK
 import arrow.mtl.extensions.writert.monad.monad
 import arrow.mtl.extensions.writert.monadWriter.monadWriter
-import arrow.core.test.UnitSpec
-import arrow.core.test.generators.GenK
-import arrow.core.test.generators.genK
-import arrow.core.test.generators.throwable
-import arrow.core.test.generators.tuple2
-import arrow.core.test.laws.AlternativeLaws
-import arrow.core.test.laws.MonadErrorLaws
-import arrow.core.test.laws.equalUnderTheLaw
 import arrow.mtl.test.MonadStateLaws
 import arrow.mtl.test.MonadTransLaws
 import arrow.mtl.test.MonadWriterLaws
