+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/EitherTTest.kt
@@ -18,6 +18,13 @@ import arrow.core.extensions.id.monad.monad
 import arrow.core.extensions.id.traverse.traverse
 import arrow.core.extensions.monoid
 import arrow.core.extensions.option.functor.functor
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.genK
+import arrow.core.test.generators.throwable
+import arrow.core.test.laws.AlternativeLaws
+import arrow.core.test.laws.DivisibleLaws
+import arrow.core.test.laws.MonadErrorLaws
+import arrow.core.test.laws.TraverseLaws
 import arrow.fx.ForIO
 import arrow.fx.IO
 import arrow.fx.extensions.io.applicative.applicative
@@ -26,6 +33,8 @@ import arrow.fx.extensions.io.functor.functor
 import arrow.fx.extensions.io.monad.monad
 import arrow.fx.mtl.concurrent
 import arrow.fx.mtl.timer
+import arrow.fx.test.eq.throwableEq
+import arrow.fx.test.laws.ConcurrentLaws
 import arrow.mtl.extensions.eithert.alternative.alternative
 import arrow.mtl.extensions.eithert.applicative.applicative
 import arrow.mtl.extensions.eithert.apply.apply
@@ -35,15 +44,8 @@ import arrow.mtl.extensions.eithert.functor.functor
 import arrow.mtl.extensions.eithert.monad.monad
 import arrow.mtl.extensions.eithert.monadError.monadError
 import arrow.mtl.extensions.eithert.traverse.traverse
-import arrow.core.test.UnitSpec
-import arrow.core.test.generators.genK
-import arrow.core.test.generators.throwable
-import arrow.core.test.laws.AlternativeLaws
-import arrow.core.test.laws.DivisibleLaws
-import arrow.core.test.laws.MonadErrorLaws
-import arrow.core.test.laws.TraverseLaws
-import arrow.core.test.laws.throwableEq
-import arrow.fx.test.laws.ConcurrentLaws
+import arrow.mtl.test.eq.eqK
+import arrow.mtl.test.generators.genK
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
 import io.kotlintest.properties.Gen
