+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/KleisliTest.kt
@@ -15,6 +15,11 @@ import arrow.core.extensions.id.monad.monad
 import arrow.core.extensions.monoid
 import arrow.core.extensions.option.alternative.alternative
 import arrow.core.extensions.option.eqK.eqK
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.GenK
+import arrow.core.test.generators.genK
+import arrow.core.test.laws.AlternativeLaws
+import arrow.core.test.laws.DivisibleLaws
 import arrow.fx.ForIO
 import arrow.fx.IO
 import arrow.fx.extensions.io.applicative.applicative
@@ -23,17 +28,14 @@ import arrow.fx.extensions.io.functor.functor
 import arrow.fx.extensions.io.monad.monad
 import arrow.fx.mtl.concurrent
 import arrow.fx.mtl.timer
+import arrow.fx.test.laws.ConcurrentLaws
 import arrow.mtl.extensions.kleisli.alternative.alternative
 import arrow.mtl.extensions.kleisli.applicative.applicative
 import arrow.mtl.extensions.kleisli.divisible.divisible
 import arrow.mtl.extensions.kleisli.functor.functor
 import arrow.mtl.extensions.kleisli.monad.monad
-import arrow.core.test.UnitSpec
-import arrow.core.test.generators.GenK
-import arrow.core.test.generators.genK
-import arrow.core.test.laws.AlternativeLaws
-import arrow.core.test.laws.DivisibleLaws
-import arrow.fx.test.laws.ConcurrentLaws
+import arrow.mtl.test.eq.eqK
+import arrow.mtl.test.generators.genK
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
 import io.kotlintest.properties.Gen
