+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/OptionTTest.kt
@@ -19,12 +19,22 @@ import arrow.core.extensions.nonemptylist.monad.monad
 import arrow.core.extensions.option.eqK.eqK
 import arrow.core.extensions.option.monad.monad
 import arrow.core.extensions.option.traverseFilter.traverseFilter
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.GenK
+import arrow.core.test.generators.genK
+import arrow.core.test.generators.option
+import arrow.core.test.laws.DivisibleLaws
+import arrow.core.test.laws.FunctorFilterLaws
+import arrow.core.test.laws.MonoidKLaws
+import arrow.core.test.laws.SemigroupKLaws
+import arrow.core.test.laws.TraverseFilterLaws
 import arrow.fx.IO
 import arrow.fx.extensions.io.concurrent.concurrent
 import arrow.fx.extensions.io.functor.functor
 import arrow.fx.extensions.io.monad.monad
 import arrow.fx.mtl.concurrent
 import arrow.fx.mtl.timer
+import arrow.fx.test.laws.ConcurrentLaws
 import arrow.mtl.extensions.ComposedFunctorFilter
 import arrow.mtl.extensions.nested
 import arrow.mtl.extensions.optiont.applicative.applicative
@@ -37,18 +47,10 @@ import arrow.mtl.extensions.optiont.monadTrans.monadTrans
 import arrow.mtl.extensions.optiont.monoidK.monoidK
 import arrow.mtl.extensions.optiont.semigroupK.semigroupK
 import arrow.mtl.extensions.optiont.traverseFilter.traverseFilter
-import arrow.core.test.UnitSpec
-import arrow.core.test.generators.GenK
-import arrow.core.test.generators.genK
-import arrow.core.test.generators.nested
-import arrow.core.test.generators.option
-import arrow.core.test.laws.DivisibleLaws
-import arrow.core.test.laws.FunctorFilterLaws
 import arrow.mtl.test.MonadTransLaws
-import arrow.core.test.laws.MonoidKLaws
-import arrow.core.test.laws.SemigroupKLaws
-import arrow.core.test.laws.TraverseFilterLaws
-import arrow.fx.test.laws.ConcurrentLaws
+import arrow.mtl.test.eq.eqK
+import arrow.mtl.test.generators.genK
+import arrow.mtl.test.generators.nested
 import arrow.typeclasses.Monad
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.forAll
