+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/StateTTests.kt
@@ -11,26 +11,28 @@ import arrow.core.extensions.option.monad.monad
 import arrow.core.extensions.option.monadCombine.monadCombine
 import arrow.core.extensions.option.semigroupK.semigroupK
 import arrow.core.extensions.tuple2.eq.eq
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.GenK
+import arrow.core.test.generators.genK
+import arrow.core.test.generators.tuple2
+import arrow.core.test.laws.MonadCombineLaws
+import arrow.core.test.laws.SemigroupKLaws
 import arrow.fx.ForIO
 import arrow.fx.IO
 import arrow.fx.extensions.io.async.async
 import arrow.fx.extensions.io.functor.functor
 import arrow.fx.extensions.io.monad.monad
 import arrow.fx.mtl.statet.async.async
+import arrow.fx.test.laws.AsyncLaws
 import arrow.mtl.extensions.statet.applicative.applicative
 import arrow.mtl.extensions.statet.functor.functor
 import arrow.mtl.extensions.statet.monad.monad
 import arrow.mtl.extensions.statet.monadCombine.monadCombine
 import arrow.mtl.extensions.statet.monadState.monadState
 import arrow.mtl.extensions.statet.semigroupK.semigroupK
-import arrow.core.test.UnitSpec
-import arrow.core.test.generators.GenK
-import arrow.core.test.generators.genK
-import arrow.core.test.generators.tuple2
-import arrow.core.test.laws.MonadCombineLaws
 import arrow.mtl.test.MonadStateLaws
-import arrow.core.test.laws.SemigroupKLaws
-import arrow.fx.test.laws.AsyncLaws
+import arrow.mtl.test.eq.eqK
+import arrow.mtl.test.generators.genK
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
 import arrow.typeclasses.Monad
