+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/test/eq/EqK.kt
@@ -1,4 +1,4 @@
-package arrow.mtl
+package arrow.mtl.test.eq
 
 import arrow.Kind
 import arrow.core.Either
@@ -10,10 +10,8 @@ import arrow.fx.extensions.io.concurrent.waitFor
 import arrow.fx.fix
 import arrow.fx.typeclasses.Duration
 import arrow.fx.typeclasses.seconds
-import arrow.core.test.generators.GenK
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
-import io.kotlintest.properties.Gen
 
 fun IO.Companion.eqK(timeout: Duration = 60.seconds) = object : EqK<ForIO> {
   override fun <A> Kind<ForIO, A>.eqK(other: Kind<ForIO, A>, EQ: Eq<A>): Boolean =
@@ -21,12 +19,5 @@ fun IO.Companion.eqK(timeout: Duration = 60.seconds) = object : EqK<ForIO> {
       IO.applicative().mapN(fix().attempt(), other.fix().attempt()) { (a, b) -> a.eqv(b) }
         .waitFor(timeout)
         .unsafeRunSync()
-    }
-}
-
-fun IO.Companion.genK() = object : GenK<ForIO> {
-  override fun <A> genK(gen: Gen<A>): Gen<Kind<ForIO, A>> =
-    gen.map {
-      IO.just(it)
     }
 }
